requirements.txt
# Core Flask dependencies
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# Database
mysql-connector-python==9.2.0
PyMySQL==1.1.1
SQLAlchemy==2.0.41

# Excel file handling
openpyxl==3.1.5
xlrd==2.0.2
xlwt==1.3.0
pyexcel==0.7.3
pyexcel-io==0.6.7
pyexcel-xls==0.7.1
pyexcel-xlsx==0.6.1
et_xmlfile==2.0.0
lml==0.2.0

# Data processing
pandas==2.2.3
numpy==2.2.4

# Environment variables
python-dotenv==1.0.1

# PDF generation
pdfkit==1.0.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.1
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
requests==2.32.3
RapidFuzz==3.13.0
Pygments==2.19.1
prompt_toolkit==3.0.50
wcwidth==0.2.13
colorama==0.4.6
psutil==7.0.0
six==1.17.0

# Compression
Brotli==1.1.0

# Security
cryptography==45.0.4
cffi==1.17.1
pycparser==2.22

# Testing
pytest==8.4.1
pluggy==1.6.0
iniconfig==2.1.0
packaging==25.0

# System dependencies (Windows only)
pywin32==310

# Other
greenlet==3.2.3
texttable==1.7.0
typing_extensions==4.14.0
tzdata==2025.1
fonttools==4.56.0